+++ src/Exception/Handler.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Hyvor\Internal\Exception;
+
+use Illuminate\Foundation\Exceptions\Handler as ExceptionHandler;
+
+class Handler extends ExceptionHandler
+{
+
+
+}