+++ src/InternalServiceProvider.php
@@ -2,9 +2,12 @@
 
 namespace Hyvor\Internal;
 
+use Hyvor\Internal\Auth\Providers\CurrentProvider;
 use Hyvor\Internal\Auth\Providers\Fake\FakeProvider;
-use Hyvor\Internal\Billing\Billing;
+use Hyvor\Internal\Auth\Providers\Hyvor\HyvorProvider;
+use Hyvor\Internal\Billing\BillingFake;
 use Hyvor\Internal\InternalApi\ComponentType;
+use Hyvor\Internal\Resource\ResourceFake;
 use Illuminate\Support\Facades\App;
 use Illuminate\Support\ServiceProvider;
 
@@ -13,7 +16,6 @@ class InternalServiceProvider extends ServiceProvider
 
     public function boot(): void
     {
-
         if (config('internal.auth.routes')) {
             $this->loadRoutesFrom(__DIR__ . '/routes/auth.php');
         }
@@ -21,6 +23,9 @@ class InternalServiceProvider extends ServiceProvider
         if (App::environment('testing')) {
             $this->loadRoutesFrom(__DIR__ . '/routes/testing.php');
         }
+
+        // set auth provider
+        CurrentProvider::set(HyvorProvider::class);
 
         if (
             config('app.env') === 'local' &&
@@ -32,7 +37,6 @@ class InternalServiceProvider extends ServiceProvider
 
     private function fake(): void
     {
-
         $class = InternalFake::class;
 
         if (class_exists('Hyvor\Internal\InternalFakeExtended')) {
@@ -42,18 +46,22 @@ class InternalServiceProvider extends ServiceProvider
         /** @var class-string<InternalFake> $class */
         $fakeConfig = new $class;
 
-        // fake user
+        // fake auth
+        FakeProvider::enable();
         $user = $fakeConfig->user();
-        config(['internal.auth.provider' => 'fake']);
         if ($user) {
             config(['internal.auth.fake.user_id' => $user->id]);
         }
         FakeProvider::databaseSet($user ? [$user] : []);
 
         // fake billing
-        Billing::fake(license: function (int $userId, ?int $resourceId, ComponentType $component) use ($fakeConfig) {
+        BillingFake::enable(license: function (int $userId, ?int $resourceId, ComponentType $component) use ($fakeConfig
+        ) {
             return $fakeConfig->license($userId, $resourceId, $component);
         });
+
+        // fake resource
+        ResourceFake::enable();
     }
 
     public function register()